Backend/requirements.txt
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.10.0
django-rest-auth==0.9.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
httplib2==0.22.0
idna==3.10
inflection==0.5.1
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
